fastapi_backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database Integration
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Django Integration
django==5.2.4
django-extensions==3.2.3
python-decouple==3.8

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Pydantic Models
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# CORS
fastapi-cors==0.0.6

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Handling
python-magic==0.4.27
aiofiles==23.2.1

# Email
aiosmtplib==3.0.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# API Documentation
fastapi-swagger-ui==0.1.0

# WebSocket Support
websockets==12.0

# Data Validation
marshmallow==3.20.1

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Environment
python-dotenv==1.0.0

# Testing
pytest-cov==4.1.0
pytest-mock==3.12.0
